Move context stuff to dedicated file and ensure `dir_fd` behavior is respected
