+++ easybuild/main.py
@@ -840,26 +840,26 @@ def prepare_main(args=None, logfile=None, testing=None):
 
 def main_with_hooks(args=None):
     # take into account that EasyBuildError may be raised when parsing the EasyBuild configuration
-    try:
-        init_session_state, eb_go, cfg_settings = prepare_main(args=args)
-    except EasyBuildError as err:
-        print_error_and_exit(err.msg, exit_code=err.exit_code)
+    # try:
+    init_session_state, eb_go, cfg_settings = prepare_main(args=args)
+    # except EasyBuildError as err:
+    #     print_error_and_exit(err.msg, exit_code=err.exit_code)
 
     hooks = load_hooks(eb_go.options.hooks)
 
-    try:
-        exit_code: EasyBuildExit = main(args=args, prepared_cfg_data=(init_session_state, eb_go, cfg_settings))
-        sys.exit(int(exit_code))
-    except EasyBuildError as err:
-        run_hook(FAIL, hooks, args=[err])
-        print_error_and_exit(err.msg, exit_code=err.exit_code)
-    except KeyboardInterrupt as err:
-        run_hook(CANCEL, hooks, args=[err])
-        print_error_and_exit("Cancelled by user: %s", err)
-    except Exception as err:
-        run_hook(CRASH, hooks, args=[err])
-        sys.stderr.write("EasyBuild crashed! Please consider reporting a bug, this should not happen...\n\n")
-        raise
+    # try:
+    exit_code: EasyBuildExit = main(args=args, prepared_cfg_data=(init_session_state, eb_go, cfg_settings))
+    sys.exit(int(exit_code))
+    # except EasyBuildError as err:
+    #     run_hook(FAIL, hooks, args=[err])
+    #     print_error_and_exit(err.msg, exit_code=err.exit_code)
+    # except KeyboardInterrupt as err:
+    #     run_hook(CANCEL, hooks, args=[err])
+    #     print_error_and_exit("Cancelled by user: %s", err)
+    # except Exception as err:
+    #     run_hook(CRASH, hooks, args=[err])
+    #     sys.stderr.write("EasyBuild crashed! Please consider reporting a bug, this should not happen...\n\n")
+    #     raise
 
 
 if __name__ == "__main__":
